Fix various bugs, mostly related to unspecified matrix sizes.

+++ src/kalman_filter.h
@@ -40,7 +40,7 @@ public:
    * using the process model
    * @param delta_T Time between k and k+1 in s
    */
-  void Predict(long long delta_T);
+  void Predict(double delta_T);
 
   /**
    * Updates the state by using standard Kalman Filter equations
